Moved fixtures to `fixtures` folder
